Concoction v0.1.0 Concoction.HTTP

Utilities for handling commmunication with the Discord API.

This will handle ratelimits, authorization & user agent headers.

Types

method()

Specs

method() :: :get | :post | :put | :patch | :delete

The HTTP method used to make the request.

Functions

init()

Specs

init() :: :ok

Initialize the tables required in ETS for the HTTP client. This storage handles information on ratelimit buckets.

maybe_float?(float)

Specs

maybe_float?(String.t()) :: float() | nil

maybe_int?(integer)

Specs

maybe_int?(String.t()) :: integer() | nil

ratelimit_remaining(url, params)

Specs

ratelimit_remaining(binary(), [any()]) :: integer()

request(method, path, data \\ %{}, params \\ [])

Make a request to the Discord API.

When a method is GET, the data parameter is passed as query-string data.

For other requests it is passed as the post body.

If a ratelimit is hit this function will block until the ratelimit is over.

Returns the status code and JSON body or nil for requests with no response.